Fix warnings

+++ Classes/MSFProductViewController.m
@@ -121,7 +121,7 @@ static NSString *const MSFAutoinputDebuggingEnvironmentKey = @"INPUT_AUTO_DEBUG"
   }];
   self.moneySlider.delegate = self;
 	RAC(self.viewModel, totalAmount) = [[self.moneySlider rac_newValueChannelWithNilValue:@0] map:^id(NSString *value) {
-		return [NSString stringWithFormat:@"%ld",value.integerValue / 100 * 100];
+		return [NSString stringWithFormat:@"%ld", (long)value.integerValue / 100 * 100];
 	 //return [NSNumber numberWithInteger:value.integerValue / 100 * 100 ];
 	}] ;
 	//self.nextPageBT.rac_command = self.viewModel.executeNextCommand;
